Reorganize restore wallet pages
